添加DeletedBills

+++ app/src/main/java/stream/com/streamapp/home/UpdateData.java
@@ -54,10 +54,14 @@ public class UpdateData {
     private static class myThread extends Thread {
         final private String url;
         final private String query;
+        final private int BillId;
+        final private int op;
 
-        myThread(String url, String query) {
+        myThread(String url, String query, int BillId, int op) {
             this.url = url;
             this.query = query;
+            this.BillId = BillId;
+            this.op = op;
         }
 
         @Override
@@ -85,7 +89,20 @@ public class UpdateData {
 //                Log.e("eee","***"+tResponse.toString());
                 String ans = tResponse.body().string();
                 result = ans;
-
+                if (op == 0)
+                {
+                    ContentValues values = new ContentValues();
+                    values.put("state", 9);
+                    DataSupport.update(Bills.class, values, BillId);
+                }
+                else if (op == 1){
+                    ContentValues values = new ContentValues();
+                    values.put("state", 9);
+                    DataSupport.update(Bills.class, values, BillId);
+                }
+                else if (op == 2){
+                    DataSupport.delete(Bills.class, BillId);
+                }
 //                Log.e("eee",ans);
 
 
@@ -110,10 +127,7 @@ public class UpdateData {
                     bills.get(i).getUser_id() + "," + bills.get(i).getAmount() + ", \"" + bills.get(i).getDate() + "\" ,\"" + bills.get(i).getPlace() + "\", \"" +
                     bills.get(i).getInOrOut() + "\",\"" + bills.get(i).getType() + "\", \"" + bills.get(i).getNote() + "\", \"" + bills.get(i).getTimeStamp() + "\", \""+bills.get(i).getMethods()+ "\" );";
             Log.d("query", query);
-            ContentValues values = new ContentValues();
-            values.put("state", 9);
-            DataSupport.update(Bills.class, values, bills.get(i).getId());
-            myThread tThread = new myThread(mUrl, query);
+            myThread tThread = new myThread(mUrl, query, bills.get(i).getId(), 0);
             tThread.start();
         }
         //修改同步
@@ -123,19 +137,17 @@ public class UpdateData {
                     bills.get(i).getPlace() + "' ,inOrOut = '" + bills.get(i).getInOrOut() + "' ,type = '" + bills.get(i).getType() + "' ,note = '"
                     + bills.get(i).getNote() + "' ,timeStamp = '" + bills.get(i).getTimeStamp() + "' ,methods = '" + bills.get(i).getMethods() + "' WHERE id = " + bills.get(i).getId() + ";";
             Log.d("query1", query);
-            ContentValues values = new ContentValues();
-            values.put("state", 9);
-            DataSupport.update(Bills.class, values, bills.get(i).getId());
-            myThread tThread = new myThread(mUrl, query);
+            myThread tThread = new myThread(mUrl, query, bills.get(i).getId(), 1);
             tThread.start();
+
         }
         //删除同步
         bills = DataSupport.where("user_id = ? and state = 3 ", String.valueOf(login.getUser_id())).find(Bills.class);
         for (int i = 0; i < bills.size(); i++) {
-            query = "DELETE FROM Bills where id = " + bills.get(i).getId() + ";";
-            Log.d("query", query);
-            DataSupport.delete(Bills.class, bills.get(i).getId());
-            myThread tThread = new myThread(mUrl, query);
+            query = "DELETE FROM Bills where id = " + bills.get(i).getId()  + " and user_id = " + bills.get(i).getUser_id() +";" + "INSERT into DeletedBills (id, user_id, timeStamp)" +
+                    "values( " + bills.get(i).getId() +"," + bills.get(i).getUser_id() + ",\"" + bills.get(i).getTimeStamp() + "\");";
+            Log.d("query2", query);
+            myThread tThread = new myThread(mUrl, query, bills.get(i).getId(), 2);
             tThread.start();
         }
 
@@ -148,14 +160,16 @@ public class UpdateData {
         Log.d("userid", "" + login.getUser_id());
         String latestTimeStamp;
         String query = "";
+        String query1 = "";
         Log.d("count", "" + count);
         if (count == 0)
             query = "select id, user_id, amount, date, place, inOrOut, note, type, timeStamp, methods from Bills where user_id = " + login.getUser_id() + ";";
         else {
             latestTimeStamp = DataSupport.max(Bills.class, "timeStamp", String.class);
             query = "select id, user_id, amount, date, place, inOrOut, note, type, timeStamp, methods from Bills where user_id = " + login.getUser_id() + " and timeStamp > '" + latestTimeStamp + "';";
+            query1 = "select id from DeletedBills where user_id = " + login.getUser_id() + "and timeStamp > '" + latestTimeStamp + "';";
         }
-        myThread tThread = new myThread(mUrl, query);
+        myThread tThread = new myThread(mUrl, query , 0 , -1);
         tThread.start();
         tThread.join();
         if (result.equals("")) {
@@ -173,88 +187,125 @@ public class UpdateData {
             for (int i = 2; i < lines.length; i++) {
                 String line = lines[i];
                 String[] bill = line.split("#");
-
-                Bills bills = new Bills();
-                bills.setId(Integer.valueOf(bill[0]));
-                bills.setUser_id(Integer.valueOf(bill[1]));
-                bills.setAmount(Double.valueOf(bill[2]));
-                bills.setDate(bill[3]);
-                bills.setPlace(bill[4]);
-                bills.setInOrOut(bill[5]);
-                bills.setNote(bill[6]);
-                bills.setType(bill[7]);
-                bills.setTimeStamp(bill[8]);
-                bills.setState(9);
-                bills.setMethods(bill[9]);
-                bills.save();
+                count = DataSupport.where("id = ?", bill[0]).count(Bills.class);
+                if (count == 0) {
+                    Bills bills = new Bills();
+                    bills.setId(Integer.valueOf(bill[0]));
+                    bills.setUser_id(Integer.valueOf(bill[1]));
+                    bills.setAmount(Double.valueOf(bill[2]));
+                    bills.setDate(bill[3]);
+                    bills.setPlace(bill[4]);
+                    bills.setInOrOut(bill[5]);
+                    bills.setNote(bill[6]);
+                    bills.setType(bill[7]);
+                    bills.setTimeStamp(bill[8]);
+                    bills.setState(9);
+                    bills.setMethods(bill[9]);
+                    bills.save();
+                }
+                else
+                {
+                    Bills bills = new Bills();
+                    bills.setUser_id(Integer.valueOf(bill[1]));
+                    bills.setAmount(Double.valueOf(bill[2]));
+                    bills.setDate(bill[3]);
+                    bills.setPlace(bill[4]);
+                    bills.setInOrOut(bill[5]);
+                    bills.setNote(bill[6]);
+                    bills.setType(bill[7]);
+                    bills.setTimeStamp(bill[8]);
+                    bills.setState(9);
+                    bills.setMethods(bill[9]);
+                    bills.update(Integer.valueOf(bill[0]));
+                }
             }
-
-            return;
         }
-    }
-
-    public static void UploadAssets() throws InterruptedException {
-        List<Assets> assets;
-        String query = "";
-        assets = DataSupport.where("user_id = ? and state = 1 ", String.valueOf(login.getUser_id())).find(Assets.class);
-        for (int i = 0; i < assets.size(); i++) {
-            query = "insert into Assets (id, user_id, amount, date,inOrOut, type, timeStamp ) values( " + assets.get(i).getId() + "," +
-                    assets.get(i).getUser_id() + "," + assets.get(i).getAmount() + ", \"" + assets.get(i).getDate() + "\", \"" +
-                    assets.get(i).getInOrOut() + "\",\"" + assets.get(i).getType() + "\", \"" + assets.get(i).getTimeStamp() + "\" );";
-            Log.d("query", query);
-            ContentValues values = new ContentValues();
-            values.put("state", 9);
-            DataSupport.update(Assets.class, values, assets.get(i).getId());
-            myThread tThread = new myThread(mUrl, query);
-            tThread.start();
+        tThread = new myThread(mUrl, query1 , 0 , -1);
+        tThread.start();
+        tThread.join();
+        if (result.equals("")) {
+            throw new internetError("网络连接错误!");
+        }
+        Log.d("result1:", result);
+        r = Pattern.compile(regex.resultPattern);
+        m = r.matcher(result);
+        if (m.find()) {
+            String res = m.group(0).replaceAll("<br>", "&");
+            res = res.replaceAll("\t", "#");
+            Log.d("res:", res);
+            String[] lines = res.split("&");
+            Log.d("length:", "" + lines.length);
+            for (int i = 2; i < lines.length; i++) {
+                String line = lines[i];
+                String[] bill = line.split("#");
+                DataSupport.delete(Bills.class, Integer.valueOf(bill[0]));
+            }
         }
         return;
     }
 
-    public static void downloadAssets() throws InterruptedException, internetError {
-        result = "";
-        int count = DataSupport.where("user_id = ?", String.valueOf(login.getUser_id())).count(Assets.class);
-        String latestTimeStamp;
-        String query = "";
-        if (count == 0)
-            query = "select id, user_id, amount, date, inOrOut, type, timeStamp from Assets where user_id = " + login.getUser_id() + ";";
-        else {
-            latestTimeStamp = DataSupport.max(Assets.class, "timeStamp", String.class);
-            query = "select id, user_id, amount, date, inOrOut, type, timeStamp from Assets where user_id = " + login.getUser_id() + " and timeStamp > '" + latestTimeStamp + "';";
-        }
-        myThread tThread = new myThread(mUrl, query);
-        tThread.start();
-        tThread.join();
+//    public static void UploadAssets() throws InterruptedException {
+//        List<Assets> assets;
+//        String query = "";
+//        assets = DataSupport.where("user_id = ? and state = 1 ", String.valueOf(login.getUser_id())).find(Assets.class);
+//        for (int i = 0; i < assets.size(); i++) {
+//            query = "insert into Assets (id, user_id, amount, date,inOrOut, type, timeStamp ) values( " + assets.get(i).getId() + "," +
+//                    assets.get(i).getUser_id() + "," + assets.get(i).getAmount() + ", \"" + assets.get(i).getDate() + "\", \"" +
+//                    assets.get(i).getInOrOut() + "\",\"" + assets.get(i).getType() + "\", \"" + assets.get(i).getTimeStamp() + "\" );";
+//            Log.d("query", query);
+//            ContentValues values = new ContentValues();
+//            values.put("state", 9);
+//            DataSupport.update(Assets.class, values, assets.get(i).getId());
+//            myThread tThread = new myThread(mUrl, query);
+//            tThread.start();
+//        }
+//        return;
+//    }
 
-        if (result.equals("")) {
-            throw new internetError("网络连接错误");
-        }
-        Log.d("result:", result);
-        Pattern r = Pattern.compile(regex.resultPattern);
-        Matcher m = r.matcher(result);
-        if (m.find()) {
-            String res = m.group(0).replaceAll("<br>", "&");
-            res = res.replaceAll("\t", "#");
-            String[] lines = res.split("&");
-            for (int i = 2; i < lines.length; i++) {
-                String line = lines[i];
-                String[] asset = line.split("#");
-
-                Assets assets = new Assets();
-                assets.setId(Integer.valueOf(asset[0]));
-                assets.setUser_id(Integer.valueOf(asset[1]));
-                assets.setAmount(Double.valueOf(asset[2]));
-                assets.setDate(asset[3]);
-                assets.setInOrOut(asset[5]);
-                assets.setType(asset[7]);
-                assets.setTimeStamp(asset[8]);
-                assets.setState(9);
-                assets.save();
-            }
-
-            return;
-        }
-    }
+//    public static void downloadAssets() throws InterruptedException, internetError {
+//        result = "";
+//        int count = DataSupport.where("user_id = ?", String.valueOf(login.getUser_id())).count(Assets.class);
+//        String latestTimeStamp;
+//        String query = "";
+//        if (count == 0)
+//            query = "select id, user_id, amount, date, inOrOut, type, timeStamp from Assets where user_id = " + login.getUser_id() + ";";
+//        else {
+//            latestTimeStamp = DataSupport.max(Assets.class, "timeStamp", String.class);
+//            query = "select id, user_id, amount, date, inOrOut, type, timeStamp from Assets where user_id = " + login.getUser_id() + " and timeStamp > '" + latestTimeStamp + "';";
+//        }
+//        myThread tThread = new myThread(mUrl, query);
+//        tThread.start();
+//        tThread.join();
+//
+//        if (result.equals("")) {
+//            throw new internetError("网络连接错误");
+//        }
+//        Log.d("result:", result);
+//        Pattern r = Pattern.compile(regex.resultPattern);
+//        Matcher m = r.matcher(result);
+//        if (m.find()) {
+//            String res = m.group(0).replaceAll("<br>", "&");
+//            res = res.replaceAll("\t", "#");
+//            String[] lines = res.split("&");
+//            for (int i = 2; i < lines.length; i++) {
+//                String line = lines[i];
+//                String[] asset = line.split("#");
+//
+//                Assets assets = new Assets();
+//                assets.setId(Integer.valueOf(asset[0]));
+//                assets.setUser_id(Integer.valueOf(asset[1]));
+//                assets.setAmount(Double.valueOf(asset[2]));
+//                assets.setDate(asset[3]);
+//                assets.setInOrOut(asset[5]);
+//                assets.setType(asset[7]);
+//                assets.setTimeStamp(asset[8]);
+//                assets.setState(9);
+//                assets.save();
+//            }
+//
+//            return;
+//        }
+//    }
 
     public static void addBills(double amount, String inOrOut, String methods) throws InterruptedException {
         Bills bill = new Bills();
